ugfs-radar/requirements.txt
# UGFS-Radar — Dépendances Python
# Pinned versions for reproducibility

# === Core / async runtime ===
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.4
structlog==24.4.0
tenacity==8.5.0
PyYAML==6.0.1

# === HTTP / scraping ===
httpx[http2]==0.27.0
selectolax==0.3.21
fake-useragent==1.5.1
playwright==1.45.0          # uniquement pour LinkedIn et sites JS-heavy
feedparser==6.0.11          # RSS feeds (EU portal, AfDB)

# === LLM / embeddings ===
groq==0.9.0                 # Llama 3.3 70B free
voyageai==0.2.3             # embeddings free tier 200M tokens
openai==1.35.0              # fallback générique (clients LLM utilisent souvent l'interface OpenAI-compatible)

# === Database ===
SQLAlchemy==2.0.31
asyncpg==0.29.0
psycopg2-binary==2.9.9      # backup sync driver
alembic==1.13.2
pgvector==0.3.2

# === Excel / Email ===
openpyxl==3.1.5
resend==1.0.1
python-dateutil==2.9.0

# === Scheduler & API ===
apscheduler==3.10.4
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
jinja2==3.1.4               # templates HTML du dashboard

# === Authentification web ===
bcrypt==4.1.3               # hashing mots de passe
PyJWT==2.8.0                # tokens session

# === Machine learning (recalibration des poids) ===
scikit-learn==1.5.0
numpy==1.26.4

# === Microsoft Graph (Teams) ===
msal==1.30.0
httpx-auth==0.22.0

# === Observability ===
sentry-sdk==2.7.1

# === Dev / test ===
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0
vcrpy==6.0.1                # mocks d'API HTTP enregistrés
ruff==0.5.0
black==24.4.2
mypy==1.10.1

# === Data analysis (pour le notebook d'exploration historique) ===
pandas==2.2.2
